Added test code for DHT11
The ADA version uses the Adafruit library and it works.  The PIE version
uses the Piettetch library and it does not work.  The latter appears to
never return from dth.acquireAndWait() and even tends to put the Photon
into an unhappy state (breathing green) from which I have to fiddle with
the buttons to even get it to OTA flash new code!