Added integrated code from Development branch to this branch
I added this code to this branch to preserve a copy of the integrated
and tested firmware before Blynk integration is added to the release
firmware in development.  It is handy to not need Blynk for hardware
development activities (PC board, packaging, cabling development).  This
firmware may be removed from this branch after the final hardware
development is completed or perhaps sooner, when Blynk integration may
be configured out if the release version firmware.