Schematic of Sensor wiring with SF RJ11 PCB
The wiring of the Sparkfun RJ11 breakout boards through a 4 wire
telephne cable is screwy, to say the least!  Here is a schematic for how
it must be wired up.  Pins 3,4, and 5 are used on the sensor side, and
pins 2,3, and 4 are used on the Box side.  Sensor 5 --> Box 2 (signal),
Sensor 4 --> Box 3 (+3.3 volt) and Sensor 3 --> Box 4 (GND).  Tested
this with > 10 feet of telephone cable and it works fine.