Update WaterLeakDetector.ino
Added DHT.begin() call to setup(), per ScruffR's suggestion.  Temp and Humidity reporting are now working!  This code is otherwise identical to the previous version, except that it no longer reports temp and humidity to the Particle Console every minute.  This latter change is necessary as a result of Particle's limiting free tier cloud publications to 100,000 per month.