WLD version 2.0
EEPROM initialization bug fixed.  There is still a latent bug that initially has a temp of -4.000 F until the first new reading.  This is visible on the servo meter - it goes max negative and then goes to the correct temp.  This wasn't an issue before temp alarms -- now sends a low alarm upon bootup, even when temp is normal.